Split PatchMgr.h
