+++ TSynth/TSynth.ino
@@ -88,6 +88,12 @@ uint8_t activeGroupIndex = 0;
 
 #include "ST7735Display.h"
 
+// Used for entering patch names.
+const static char CHARACTERS[TOTALCHARS] = {'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',' ', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z', ' ', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0'};
+int charIndex;
+char currentCharacter;
+String renamedPatch;
+
 //USB HOST MIDI Class Compliant
 USBHost myusb;
 USBHub hub1(myusb);
